refactor(ricer::data_xchg)!: rename `config` module to `data_xchg` module
The name `data_xchg` is more accurate than `config` for the purpose of managing
the TOML data exchange format in Ricer.
