+++ src/tests/data_xchg/cmd_hook.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 Jason Pena <jasonpena@awkless.com>
 // SPDX-License-Identifier: MIT
 
-use crate::config::{CommandHook, ConfigEntry, Hook};
+use crate::data_xchg::{CommandHook, Hook, TomlEntry};
 
 use anyhow::Result;
 use indoc::indoc;
